Move key states from geoapp to separate files
