+++ Makefile
@@ -1,6 +1,6 @@
 CFLAGS = -g -Wshadow -std=c++11 -pedantic -Wall -Wextra -Wcast-align -Wcast-qual -Wctor-dtor-privacy -Winit-self -Wlogical-op -Wmissing-declarations -Wmissing-include-dirs -Wnoexcept  -Woverloaded-virtual -Wredundant-decls -Wsign-conversion -Wsign-promo -Wstrict-null-sentinel -Wstrict-overflow=5 -Wswitch-default -Wundef -Wno-unused -Wold-style-cast -Wformat-nonliteral -Wpointer-arith -Winline -Wwrite-strings -Wno-unused-parameter -Wfloat-equal -fsanitize=undefined -fsanitize=unreachable -fsanitize=leak -fsanitize=enum -fsanitize=builtin -fsanitize=shift-exponent -fsanitize=pointer-compare -fsanitize=pointer-subtract -fsanitize=address -Werror
 
-OBJECTS = obj/main.o obj/Construction.o obj/drawers.o obj/Geoapp.o obj/geo.o obj/geoImpl.o obj/Input.o obj/inputSFML.o
+OBJECTS = obj/main.o obj/Construction.o obj/drawers.o obj/Geoapp.o obj/geo.o obj/geoImpl.o obj/Input.o obj/inputSFML.o obj/keyStates.o
 
 bin/app: $(OBJECTS)
 	g++ $(OBJECTS) -o bin/app -lsfml-window -lsfml-system -lsfml-graphics $(CFLAGS)
@@ -20,7 +20,7 @@ obj/drawers.o: source/drawers.cpp source/drawers.h source/geo.h | obj
 obj/Geoapp.o: source/Geoapp.cpp source/Geoapp.h source/geo.h \
  source/inputSFML.h source/Input.h source/Construction.h \
  source/vectorHolder.h source/drawers.h source/uiOptions.h \
- source/conditionTracker.h | obj
+ source/conditionTracker.h source/keyStates.h | obj
 	g++ source/Geoapp.cpp -o obj/Geoapp.o -c $(CFLAGS)
 
 obj/geo.o: source/geo.cpp source/geo.h | obj
@@ -34,6 +34,12 @@ obj/Input.o: source/Input.cpp source/Input.h | obj
 
 obj/inputSFML.o: source/inputSFML.cpp source/inputSFML.h source/Input.h | obj
 	g++ source/inputSFML.cpp -o obj/inputSFML.o -c $(CFLAGS)
+
+obj/keyStates.o: source/keyStates.cpp source/keyStates.h source/Geoapp.h \
+ source/geo.h source/inputSFML.h source/Input.h source/Construction.h \
+ source/vectorHolder.h source/drawers.h source/uiOptions.h \
+ source/conditionTracker.h | obj
+	g++ source/keyStates.cpp -o obj/keyStates.o -c $(CFLAGS)
 
 obj:
 	mkdir obj
