+++ source/Geoapp.h
@@ -112,129 +112,3 @@ class Geoapp{
             currentMode = m;
         }
 };
-
-
-class inputCameraMovementState : public inputState {
-    private:
-        Geoapp* app;
-        double x, y;
-    public:
-        inputCameraMovementState (inputManager* _manager, Geoapp* _app, double _x, double _y) : inputState(_manager), app(_app), x(_x), y(_y) {}
-        virtual void onEnter () override {
-            app->moveCamera (x,y);
-            done();
-        }
-};
-
-class inputPointMovementState : public inputState {
-    private:
-        Geoapp* app;
-        double x, y;
-    public:
-        inputPointMovementState (inputManager* _manager, Geoapp* _app, double _x, double _y) : inputState (_manager), app(_app), x(_x), y(_y) {}
-        virtual void onEnter () override {
-            app->moveHulledPoints(x,y);
-            done();
-        }
-};
-
-class inputPointSelectionState : public inputState {
-    private:
-        Geoapp* app;
-    public:
-        inputPointSelectionState (inputManager* _manager, Geoapp* _app) :inputState(_manager), app(_app) {}
-        virtual void onEnter () override {
-            app->setCurrentMode (Geoapp::mode::selection);
-            done();
-        }
-};
-
-class inputPointCreationState : public inputState {
-    private:
-        Geoapp* app;
-    public:
-        inputPointCreationState (inputManager* _manager, Geoapp* _app) :inputState(_manager), app(_app) {}
-        virtual void onEnter () override {
-            app->setCurrentMode (Geoapp::mode::pointCreation);
-            done();
-        }
-};
-
-class inputSetMarkState : public inputState {
-    private:
-        Geoapp* app;
-    public:
-        inputSetMarkState (inputManager* _manager, Geoapp* _app) : inputState(_manager), app(_app) {}
-        virtual void onKey (inputManager::keyType k, inputManager::action a, unsigned int mods){
-            if (a == inputManager::released) return;
-            char mk = inputManager::keyToChar(k);
-            if (mk == 0) {
-                done();
-                return;
-            }
-            app->setMark (mk,app->getCamera());
-            done();
-        }
-};
-
-class inputGoToMarkState : public inputState {
-    private:
-        Geoapp* app;
-    public:
-        inputGoToMarkState (inputManager* _manager, Geoapp* _app) : inputState(_manager), app(_app) {}
-        virtual void onKey (inputManager::keyType k, inputManager::action a, unsigned int mods){
-            if (a == inputManager::released) return;
-            char mk = inputManager::keyToChar(k);
-            if (mk == 0) {
-                done();
-                return;
-            }
-            const Point * const pptr = app->getMark (mk);
-            if (pptr != NULL) {
-                app->setCamera (*pptr);
-            }
-            done();
-        }
-};
-
-class inputUIScrollState : public inputState {
-    private:
-        Geoapp* app;
-        double scrollValue;
-    public:
-        inputUIScrollState (inputManager* _manager, Geoapp* _app, double _scrollValue) : inputState(_manager), app(_app), scrollValue (_scrollValue){}
-        virtual void onEnter () override {
-            app->scrollUI (scrollValue);
-            done();
-        }
-};
-
-class inputScalingState : public inputState {
-    private:
-        Geoapp* app;
-        double scaleValue;
-    public:
-        inputScalingState (inputManager* _manager, Geoapp* _app, double _scaleValue) : inputState (_manager), app(_app), scaleValue (_scaleValue) {}
-        virtual void onEnter () override {
-            app->changeScale (scaleValue);
-            done();
-        }
-};
-
-class inputSaveState : public inputState {
-    private:
-        Geoapp* app;
-        const std::string filename;
-        svgDrawingClass* drawer;
-    public:
-        inputSaveState (inputManager* _manager, Geoapp* _app, const std::string& _filename, svgDrawingClass* _drawer) :inputState (_manager), app(_app), filename(_filename), drawer(_drawer) {}
-        virtual void onEnter () override {
-            drawer->setVisible (app->getVisible());
-            drawer->setBox (app->getBox());
-            drawer->startDrawing();
-            app->drawShapes (drawer);
-            drawer->endDrawing();
-            drawer->saveToFile (filename);
-            done();
-        }
-};
