+++ source/Input.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <map>
 #include <cstddef>
